Fixed program test
