+++ tests/program/test_validate.py
@@ -240,7 +240,6 @@ def test_program_validate_long_error_messages_976():
     assert output.strip() == expected.strip()
 
 
-@pytest.mark.skip
 def test_program_validate_partial_validation_info_933():
     result = runner.invoke(program, "validate data/countries.csv --limit-errors 2")
     assert result.exit_code == 1
@@ -259,7 +258,6 @@ def test_program_validate_summary_1094():
     assert result.stdout.count("Total Time Taken (sec)")
 
 
-@pytest.mark.skip
 def test_program_validate_single_resource_221():
     result = runner.invoke(
         program, "validate data/datapackage.json --resource-name number-two"
