--- tests/test_errors.py
@@ -1,9 +1,0 @@
-from frictionless import errors
-
-
-# General
-
-
-def test_error_from_exception():
-    error = errors.SourceError(note="note")
-    assert error["code"] == "source-error"
